+++ legacy/WUI/Table/WUITable-0.2.js
@@ -1,12 +1,12 @@
 /*
- * WUITable - v0.1
+ * WUITable - v0.2
  * Author: Sergio E. Belmar (wuijs.project@gmail.com)
  * Copyright (c) Sergio E. Belmar (wuijs.project@gmail.com)
  */
 
 class WUITable {
 
-	static version = "0.1";
+	static version = "0.2";
 	static #defaults = {
 		selector: ".wui-table",
 		width: "auto",